Odev1son
Define a 15 x 15 two-dimensional character array. Initialize that array with the following characters: {'G','S','L','N','F','R','N','P','R','E','N','D','P','O','R'},
                       {'N','A','N','F','D','K','U','O','L','M','K','I','A','C','N'},
                       {'I','F','V','A','C','S','F','P','T','A','A','S','R','O','H'},
                       {'L','E','P','I','I','S','I','N','R','S','T','N','T','A','A'},
                       {'I','S','R','L','I','R','L','N','M','I','Y','I','D','Y','D'},
                       {'A','B','T','L','A','I','D','G','S','O','S','R','T','E','R'},
                       {'W','U','E','N','G','T','N','A','F','I','S','I','D','G','R'},
                       {'G','N','A','H','E','I','N','S','H','S','V','S','M','O','O'},
                       {'C','H','T','O','N','V','I','E','C','A','P','S','S','T','B'},
                       {'L','O','A','I','U','B','R','A','C','K','E','T','E','O','R'},
                       {'W','I','A','D','E','T','N','U','O','M','R','R','L','T','O'},
                       {'R','T','G','E','R','M','S','E','U','E','I','E','I','H','W'},
                       {'E','F','S','H','A','D','R','I','E','F','E','P','T','E','N'},
                       {'A','S','W','I','T','C','H','T','D','P','L','A','Q','U','E'},
                       {'D','N','E','R','G','S','H','E','L','E','S','P','A','C','N'}
Read a word from user and search that word in the two-dimensional array.  If your program finds the word in the array, it prints all the elements of the array with the following rule. The word will be printed as it is.  The other characters will be printed as ‘*’. If your word is not in the array, all the characters will be printed with ‘*’ character. The words should be searched from left to right and from top to bottom.